Add Webpack DLL during development
This massively speeds up recompiling and hot reloading.

The only downside is the DLL must be built before starting
the webpack dev server for the first time, and again after
updating installed modules.

The initial list of modules included in the DLL is based on
the largest modules in the browser bundle according to
the output of the Webpack bundle analysis.